[KiMage-10] Detach event bus tests from implementation
